+++ h2integrate/converters/natural_gas/natural_gas_cc_ct.py
@@ -114,6 +114,14 @@ class NaturalGasPerformanceModel(PerformanceModelBaseClass):
             desc="Natural gas input energy",
         )
 
+        self.add_output(
+            "unmet_electricity_demand",
+            val=0.0,
+            shape=n_timesteps,
+            units=self.commodity_rate_units,
+            desc="Unmet electricity demand for natural gas plant",
+        )
+
     def compute(self, inputs, outputs):
         """
         Compute electricity output from natural gas input.
@@ -125,7 +133,8 @@ class NaturalGasPerformanceModel(PerformanceModelBaseClass):
         Args:
             inputs: OpenMDAO inputs object containing natural_gas_in, heat_rate_mmbtu_per_mwh,
                 system_capacity, and electricity_demand.
-            outputs: OpenMDAO outputs object for electricity_out and natural_gas_consumed
+            outputs: OpenMDAO outputs object for electricity_out, natural_gas_consumed,
+                and unmet_electricity_demand.
         """
 
         # calculate max input and output
@@ -166,6 +175,7 @@ class NaturalGasPerformanceModel(PerformanceModelBaseClass):
         outputs["annual_electricity_produced"] = outputs["total_electricity_produced"] * (
             1 / self.fraction_of_year_simulated
         )
+        outputs["unmet_electricity_demand"] = inputs["electricity_demand"] - electricity_out
 
 
 @define(kw_only=True)