+++ h2integrate/core/supported_models.py
@@ -103,6 +103,10 @@ from h2integrate.converters.natural_gas.natural_gas_cc_ct import (
 )
 from h2integrate.converters.hydrogen.singlitico_cost_model import SingliticoCostModel
 from h2integrate.converters.co2.marine.direct_ocean_capture import DOCCostModel, DOCPerformanceModel
+from h2integrate.converters.hydrogen.steam_methane_reformer import (
+    SteamMethaneReformerCostModel,
+    SteamMethaneReformerPerformanceModel,
+)
 from h2integrate.converters.hydrogen.geologic.mathur_modified import GeoH2SubsurfaceCostModel
 from h2integrate.resource.solar.nlr_developer_goes_api_models import (
     GOESTMYSolarAPI,
@@ -205,6 +209,8 @@ supported_models = {
     "WOMBATElectrolyzerModel": WOMBATElectrolyzerModel,
     "LinearH2FuelCellPerformanceModel": LinearH2FuelCellPerformanceModel,
     "H2FuelCellCostModel": H2FuelCellCostModel,
+    "SteamMethaneReformerPerformanceModel": SteamMethaneReformerPerformanceModel,
+    "SteamMethaneReformerCostModel": SteamMethaneReformerCostModel,
     "SimpleASUCostModel": SimpleASUCostModel,
     "SimpleASUPerformanceModel": SimpleASUPerformanceModel,
     "HOPPComponent": HOPPComponent,